+++ tmsgit/__main__.py
@@ -27,6 +27,10 @@ def register_fix_parser(subparsers):
     create_parser = subparsers.add_parser('fix', help='Rewrite last commit to integrate current changes.')
 
 
+def register_init_parser(subparsers):
+    create_parser = subparsers.add_parser('init', help='Create default configuration files in repository.')
+
+
 def parse_args():
     parser = argparse.ArgumentParser()
     subparsers = parser.add_subparsers(dest='command')
@@ -37,6 +41,7 @@ def parse_args():
     register_commit_parser(subparsers)
     register_log_parser(subparsers)
     register_fix_parser(subparsers)
+    register_init_parser(subparsers)
 
     argcomplete.autocomplete(parser)
 
@@ -69,3 +74,6 @@ def main():
     elif args.command == 'fix':
         from .commands import fix_previous_commit
         fix_previous_commit()
+    elif args.command == 'init':
+        from .commands import initialize_project
+        initialize_project()